+++ src/main/java/com/epam/songservice/repository/SongMetadataRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SongMetadataRepository extends JpaRepository<SongMetadata, Long> {
+public interface SongMetadataRepository extends JpaRepository<SongMetadata, Integer> {
 
-    Optional<SongMetadata> findByResourceId(Long id);
+    Optional<SongMetadata> findByResourceId(Integer id);
 }
